# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChittyChronicle API",
    "version": "1.0.0",
    "description": "Legal timeline management system with AI-powered contradiction detection, evidence tracking, and ChittyOS ecosystem integration.",
    "contact": {
      "name": "ChittyOS Support",
      "url": "https://chitty.cc"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://chronicle.chitty.cc",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "description": "Returns the authenticated user's profile with ChittyID information",
        "security": [{"cookieAuth": []}],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session cookie from ChittyID or Replit Auth"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ChittyID or user ID"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              }
            }
          },
          "profile": {
            "type": "object",
            "description": "ChittyID profile with attestations"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}